ai_code/requirements.txt
beautifulsoup4
chromadb
crewai
ddgs
duckduckgo-search
fpdf
google
graphviz
langchain
langchain-chroma
langchain-google-genai
langchain-openai
langchain_community
langchain_core
langchain_groq
langchain_huggingface
langgraph
markdown
numpy
pandas
pinecone
plotly
pydantic
pypdf
python-dotenv
reportlab
sentence-transformers
streamlit
Wikipedia-API
youtube-transcript-api